+++ test/complex/compound/arrays.js
@@ -3,8 +3,8 @@ import {run} from '../../runner.js';
 describe('Valid compound arrays:', function() {
   describe('already-supported combinations', function() {
     run([
-      [['30', '9', '*', '*', 'MON-FRI'], 'every Monday-Friday at 9:30 AM'],
-      [['0', '22', '*', '*', '1-5'], 'every Monday-Friday at 10:00 PM'],
+      [['30', '9', '*', '*', 'MON-FRI'], 'every Monday through Friday at 9:30 AM'],
+      [['0', '22', '*', '*', '1-5'], 'every Monday through Friday at 10:00 PM'],
       [['0', '0', '25', '12', '*'], 'on December 25th at 12:00 AM']
     ]);
   });
@@ -20,7 +20,7 @@ describe('Valid compound arrays:', function() {
     run([
       [['0,30', '9', '*', '*', '*'], 'every day at 9:00 AM and 9:30 AM'],
       [['0,30', '9', '*', '*', 'MON-FRI'],
-        'every Monday-Friday at 9:00 AM and 9:30 AM'],
+        'every Monday through Friday at 9:00 AM and 9:30 AM'],
       [['0,30', '9,17', '*', '*', '*'],
         'every day at 9:00 AM, 9:30 AM, 5:00 PM and 5:30 PM']
     ]);
@@ -31,7 +31,7 @@ describe('Valid compound arrays:', function() {
       [['*/15', '9-17', '*', '*', '*'],
         'every 15 minutes from 9:00 AM through 5:45 PM'],
       [['*/15', '9-17', '*', '*', 'MON-FRI'],
-        'every 15 minutes from 9:00 AM through 5:45 PM on Monday-Friday']
+        'every 15 minutes from 9:00 AM through 5:45 PM on Monday through Friday']
     ]);
   });
 
@@ -45,7 +45,7 @@ describe('Valid compound arrays:', function() {
         'at zero and 30 minutes past the hour from 9:00 AM through 5:30 PM'],
       [['15', '9-17', '*', '*', 'MON-FRI'],
         'at 15 minutes past the hour from 9:00 AM through 5:15 PM ' +
-          'on Monday-Friday']
+          'on Monday through Friday']
     ]);
   });
 
@@ -60,7 +60,7 @@ describe('Valid compound arrays:', function() {
     run([
       [['*', '*', '*', '*', 'MON'], 'every minute on Monday'],
       [['0', '*', '*', '*', 'MON'], 'every hour on Monday'],
-      [['*', '*', '*', '*', 'MON-FRI'], 'every minute on Monday-Friday'],
+      [['*', '*', '*', '*', 'MON-FRI'], 'every minute on Monday through Friday'],
       [['*', '*', '13', '*', '*'], 'every minute on the 13th'],
       [['0', '*', '13', '*', '*'], 'every hour on the 13th'],
       [['0', '*', '*', '1', '*'], 'every hour in January'],